[yul-phaser] Refactor the code from main into a Phaser class
