+++ tools/CMakeLists.txt
@@ -15,6 +15,8 @@ install(TARGETS solidity-upgrade DESTINATION "${CMAKE_INSTALL_BINDIR}")
 
 add_executable(yul-phaser
 	yulPhaser/main.cpp
+	yulPhaser/Phaser.h
+	yulPhaser/Phaser.cpp
 	yulPhaser/GeneticAlgorithms.h
 	yulPhaser/GeneticAlgorithms.cpp
 	yulPhaser/Population.h
